+++ kubernetes-model/kubernetes-model-core/cmd/generate/generate.go
@@ -118,7 +118,7 @@ func main() {
     reflect.TypeOf(time.Time{}): reflect.TypeOf(""),
     reflect.TypeOf(struct{}{}):  reflect.TypeOf(""),
   }
-  schema, err := schemagen.GenerateSchema(reflect.TypeOf(Schema{}), packages, typeMap, customTypeNames, "kubernetes")
+  schema, err := schemagen.GenerateSchema(reflect.TypeOf(Schema{}), packages, typeMap, customTypeNames, "core")
   if err != nil {
     fmt.Fprintf(os.Stderr, "An error occurred: %v", err)
     return